Top charts: Correctly extract app prices
This results in a breaking change (examples are for en_DE, other
locales will differ):

Previously, the price of free apps was extracted as `0 EUR` and the
price for paid apps was (incorrectly) extracted e.g. as
`4990000 EUR €4.99`.

Now, the price is also correctly extracted for paid apps and
consistently formatted for both using `Intl.NumberFormat`. This is
necessary because the `batchexecute` endpoint returns a formatted
price string for paid apps but not for free ones.